What - make kpfs_curl_range_get be introspected
@Description/root cause - Why
use off_t and "-D_FILE_OFFSET_BITS=64" to let kpfs_curl_range_get be introspected.

@Resolution - How
1. use off_t to replace "long long" which could not let kpfs_curl_range_get be introspected.
2. add -D_FILE_OFFSET_BITS=64 in CFLAGS, so that off_t will be "long long int"

@Reviewer
None

@Tickets
NONE

+++ trunk/src/kpfs_curl.h
@@ -24,6 +24,6 @@
 #ifndef	__KPFS_CURL_H__
 #define	__KPFS_CURL_H__
 
-int kpfs_curl_range_get(const char *url, char *buf, long long start_pos, long long end_pos);
+int kpfs_curl_range_get(const char *url, char *buf, off_t start_pos, off_t end_pos);
 int kpfs_curl_upload(const char *url, char *file, char *reply);
 #endif
